% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_data.R
\docType{data}
\name{rand_data}
\alias{rand_data}
\title{Rand_data}
\format{
A vector with 100 entries drawn from a normal distribution N(4,121) I believe.
}
\usage{
rand_data
}
\description{
Rand_data
}
\keyword{datasets}
