% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{Standardize a vector}
\usage{
standardize(data)
}
\arguments{
\item{data}{A numerical vector to be standardized.}
}
\value{
A vector that is the standardized version of data.
}
\description{
Standardize a vector
}
\examples{
x <- c(1,0,1,3,3)
standardize(x)

}
